requirements.txt
flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.1
langchain==0.3.7
langchain-openai==0.2.8
langchain-core==0.3.15

# Testing dependencies
pytest==8.3.3
pytest-cov==5.0.0

# Optional development dependencies
gunicorn==23.0.0
waitress==3.0.0
